pyproject.toml
[tool.poetry]
name = "infrahub_ansible_modules"
version = "0.0.1"
description = "Ansible collection to interact with Infrahub's API"
authors = ["Benoit Kohler <benoit@opsmill.com"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
ansible-core = "^2.14"
black = "*"
codecov = "*"
coverage = "^6.5"
deepdiff = "*"
cryptography = "*"
jinja2 = "*"
jmespath = "*"
pynetbox = "^7"
pytest = "*"
pytest-mock = "*"
pytest-xdist = "*"
pytest-forked = "^1.6.0"
pyyaml = "*"
mock = "^4.0.2"
antsibull = "^0.45.1"
importlib-metadata = "1.7.0"
pylint = "^2.6.0"
pytz = "*"
packaging = "*"
jsondiff = "*"
sphinx_rtd_theme = "*"
hypothesis = "^6.8.0"
pynautobot = "^1.1.2"
pytest-pythonpath = "^0.7.3"
parameterized = "^0.8.1"
invoke = "^1.6.0"
bandit = "^1.7.0"
antsibull-docs = "^1.7.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]

select = [
    "C90",
    "DTZ",
    "E",
    "F",
    "TCH",
    "W",
]

line-length = 170

[tool.ruff.mccabe]
# Target max-complexity=10
max-complexity = 26

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
python_paths = "./"
addopts = "-vv"